refactor manifest-writer
