+++ iceberg-rust-spec/src/spec/manifest.rs
@@ -1,182 +1,24 @@
 /*!
 Manifest files
 */
-use std::{
-    collections::HashMap,
-    io::Read,
-    iter::{repeat, Map, Repeat, Zip},
-    ops::{Deref, DerefMut},
-    sync::Arc,
-};
+use std::collections::HashMap;
 
-use apache_avro::{
-    types::Value as AvroValue, Reader as AvroReader, Schema as AvroSchema, Writer as AvroWriter,
-};
+use apache_avro::Schema as AvroSchema;
 use derive_builder::Builder;
 use derive_getters::Getters;
 use serde::{de::DeserializeOwned, ser::SerializeSeq, Deserialize, Serialize};
 use serde_bytes::ByteBuf;
 use serde_repr::{Deserialize_repr, Serialize_repr};
 
-use crate::{
-    error::Error,
-    spec::schema::{SchemaV1, SchemaV2},
-};
+use crate::error::Error;
 
 use super::{
     partition::{PartitionField, PartitionSpec},
     schema::Schema,
-    table_metadata::{FormatVersion, TableMetadata},
+    table_metadata::FormatVersion,
     types::{PrimitiveType, StructType, Type},
     values::{Struct, Value},
 };
-
-type ReaderZip<'a, R> = Zip<AvroReader<'a, R>, Repeat<Arc<(Schema, PartitionSpec, FormatVersion)>>>;
-type ReaderMap<'a, R> = Map<
-    ReaderZip<'a, R>,
-    fn(
-        (
-            Result<AvroValue, apache_avro::Error>,
-            Arc<(Schema, PartitionSpec, FormatVersion)>,
-        ),
-    ) -> Result<ManifestEntry, Error>,
->;
-
-/// Iterator of ManifestFileEntries
-pub struct ManifestReader<'a, R: Read> {
-    reader: ReaderMap<'a, R>,
-}
-
-impl<'a, R: Read> Iterator for ManifestReader<'a, R> {
-    type Item = Result<ManifestEntry, Error>;
-    fn next(&mut self) -> Option<Self::Item> {
-        self.reader.next()
-    }
-}
-
-impl<'a, R: Read> ManifestReader<'a, R> {
-    /// Create a new ManifestFile reader
-    pub fn new(reader: R) -> Result<Self, Error> {
-        let reader = AvroReader::new(reader)?;
-        let metadata = reader.user_metadata();
-
-        let format_version: FormatVersion = match metadata
-            .get("format-version")
-            .map(|bytes| String::from_utf8(bytes.clone()))
-            .transpose()?
-            .unwrap_or("1".to_string())
-            .as_str()
-        {
-            "1" => Ok(FormatVersion::V1),
-            "2" => Ok(FormatVersion::V2),
-            _ => Err(Error::InvalidFormat("format version".to_string())),
-        }?;
-
-        let schema: Schema = match format_version {
-            FormatVersion::V1 => TryFrom::<SchemaV1>::try_from(serde_json::from_slice(
-                metadata
-                    .get("schema")
-                    .ok_or(Error::InvalidFormat("manifest metadata".to_string()))?,
-            )?)?,
-            FormatVersion::V2 => TryFrom::<SchemaV2>::try_from(serde_json::from_slice(
-                metadata
-                    .get("schema")
-                    .ok_or(Error::InvalidFormat("manifest metadata".to_string()))?,
-            )?)?,
-        };
-
-        let partition_fields: Vec<PartitionField> = serde_json::from_slice(
-            metadata
-                .get("partition-spec")
-                .ok_or(Error::InvalidFormat("manifest metadata".to_string()))?,
-        )?;
-        let spec_id: i32 = metadata
-            .get("partition-spec-id")
-            .map(|x| String::from_utf8(x.clone()))
-            .transpose()?
-            .unwrap_or("0".to_string())
-            .parse()?;
-        let partition_spec = PartitionSpec::builder()
-            .with_spec_id(spec_id)
-            .with_fields(partition_fields)
-            .build()?;
-        Ok(Self {
-            reader: reader
-                .zip(repeat(Arc::new((schema, partition_spec, format_version))))
-                .map(avro_value_to_manifest_entry),
-        })
-    }
-}
-
-/// A writer for manifest entries
-pub struct ManifestWriter<'a, W: std::io::Write>(AvroWriter<'a, W>);
-
-impl<'a, W: std::io::Write> Deref for ManifestWriter<'a, W> {
-    type Target = AvroWriter<'a, W>;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl<'a, W: std::io::Write> DerefMut for ManifestWriter<'a, W> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-
-impl<'a, W: std::io::Write> ManifestWriter<'a, W> {
-    pub fn new(
-        writer: W,
-        schema: &'a AvroSchema,
-        table_metadata: &TableMetadata,
-        branch: Option<&str>,
-    ) -> Result<Self, Error> {
-        let mut avro_writer = AvroWriter::new(schema, writer);
-
-        avro_writer.add_user_metadata(
-            "format-version".to_string(),
-            match table_metadata.format_version {
-                FormatVersion::V1 => "1".as_bytes(),
-                FormatVersion::V2 => "2".as_bytes(),
-            },
-        )?;
-
-        avro_writer.add_user_metadata(
-            "schema".to_string(),
-            match table_metadata.format_version {
-                FormatVersion::V1 => serde_json::to_string(&Into::<SchemaV1>::into(
-                    table_metadata.current_schema(branch)?.clone(),
-                ))?,
-                FormatVersion::V2 => serde_json::to_string(&Into::<SchemaV2>::into(
-                    table_metadata.current_schema(branch)?.clone(),
-                ))?,
-            },
-        )?;
-
-        avro_writer.add_user_metadata(
-            "schema-id".to_string(),
-            serde_json::to_string(&table_metadata.current_schema(branch)?.schema_id())?,
-        )?;
-
-        avro_writer.add_user_metadata(
-            "partition-spec".to_string(),
-            serde_json::to_string(&table_metadata.default_partition_spec()?.fields())?,
-        )?;
-
-        avro_writer.add_user_metadata(
-            "partition-spec-id".to_string(),
-            serde_json::to_string(&table_metadata.default_partition_spec()?.spec_id())?,
-        )?;
-
-        avro_writer.add_user_metadata("content".to_string(), "data")?;
-
-        Ok(ManifestWriter(avro_writer))
-    }
-
-    pub fn into_inner(self) -> Result<W, Error> {
-        Ok(self.0.into_inner()?)
-    }
-}
 
 /// Entry in manifest with the iceberg spec version 2.
 #[derive(Debug, Serialize, PartialEq, Clone, Getters, Builder)]
@@ -205,7 +47,7 @@ impl ManifestEntry {
 }
 
 impl ManifestEntry {
-    pub(crate) fn try_from_v2(
+    pub fn try_from_v2(
         value: ManifestEntryV2,
         schema: &Schema,
         partition_spec: &PartitionSpec,
@@ -219,7 +61,7 @@ impl ManifestEntry {
         })
     }
 
-    pub(crate) fn try_from_v1(
+    pub fn try_from_v1(
         value: ManifestEntryV1,
         schema: &Schema,
         partition_spec: &PartitionSpec,
@@ -1480,32 +1322,6 @@ impl DataFileV2 {
                 }
             ]
         }"#
-    }
-}
-
-#[allow(clippy::type_complexity)]
-// Convert avro value to ManifestEntry based on the format version of the table.
-fn avro_value_to_manifest_entry(
-    value: (
-        Result<AvroValue, apache_avro::Error>,
-        Arc<(Schema, PartitionSpec, FormatVersion)>,
-    ),
-) -> Result<ManifestEntry, Error> {
-    let entry = value.0?;
-    let schema = &value.1 .0;
-    let partition_spec = &value.1 .1;
-    let format_version = &value.1 .2;
-    match format_version {
-        FormatVersion::V2 => ManifestEntry::try_from_v2(
-            apache_avro::from_value::<ManifestEntryV2>(&entry)?,
-            schema,
-            partition_spec,
-        ),
-        FormatVersion::V1 => ManifestEntry::try_from_v1(
-            apache_avro::from_value::<ManifestEntryV1>(&entry)?,
-            schema,
-            partition_spec,
-        ),
     }
 }
