+++ iceberg-rust/src/table/mod.rs
@@ -4,6 +4,7 @@ Defining the [Table] struct that represents an iceberg table.
 
 use std::{io::Cursor, iter::repeat, sync::Arc};
 
+use manifest::ManifestReader;
 use object_store::{path::Path, ObjectStore};
 
 use futures::{
@@ -12,7 +13,7 @@ use futures::{
 use iceberg_rust_spec::util::{self};
 use iceberg_rust_spec::{
     spec::{
-        manifest::{Content, ManifestEntry, ManifestReader},
+        manifest::{Content, ManifestEntry},
         manifest_list::ManifestListEntry,
         schema::Schema,
         table_metadata::TableMetadata,
