+++ iceberg-rust/src/table/transaction/operation.rs
@@ -8,24 +8,19 @@ use iceberg_rust_spec::manifest_list::{
     manifest_list_schema_v1, manifest_list_schema_v2, ManifestListReader,
 };
 use iceberg_rust_spec::spec::table_metadata::TableMetadata;
-use iceberg_rust_spec::table_metadata::FormatVersion;
-use iceberg_rust_spec::util::strip_prefix;
-use iceberg_rust_spec::{
-    manifest::ManifestReader,
-    spec::{
-        manifest::{partition_value_schema, DataFile, ManifestEntry, ManifestWriter, Status},
-        manifest_list::{Content, FieldSummary, ManifestListEntry},
-        partition::PartitionField,
-        schema::Schema,
-        snapshot::{
-            generate_snapshot_id, SnapshotBuilder, SnapshotReference, SnapshotRetention, Summary,
-        },
-        values::{Struct, Value},
+use iceberg_rust_spec::spec::{
+    manifest::{partition_value_schema, DataFile, ManifestEntry, Status},
+    schema::Schema,
+    snapshot::{
+        generate_snapshot_id, SnapshotBuilder, SnapshotReference, SnapshotRetention, Summary,
     },
 };
+use iceberg_rust_spec::table_metadata::FormatVersion;
+use iceberg_rust_spec::util::strip_prefix;
 use object_store::ObjectStore;
 use smallvec::SmallVec;
 
+use crate::table::manifest::{ManifestReader, ManifestWriter};
 use crate::{
     catalog::commit::{TableRequirement, TableUpdate},
     error::Error,
@@ -216,16 +211,7 @@ impl Operation {
                 // Write manifest files
                 // Split manifest file if limit is exceeded
                 if n_splits == 0 {
-                    // If manifest doesn't need to be split
-                    let mut manifest_writer = ManifestWriter::new(
-                        Vec::new(),
-                        &manifest_schema,
-                        table_metadata,
-                        branch.as_deref(),
-                    )?;
-
-                    // Copy potential existing entries
-                    if let Some(manifest) = &manifest {
+                    let mut manifest_writer = if let Some(manifest) = manifest {
                         let manifest_bytes: Vec<u8> = object_store
                             .get(&strip_prefix(&manifest.manifest_path).as_str().into())
                             .await?
@@ -233,12 +219,14 @@ impl Operation {
                             .await?
                             .into();
 
-                        let manifest_reader = apache_avro::Reader::new(&*manifest_bytes)?;
-                        manifest_writer.extend(manifest_reader.filter_map(Result::ok))?;
-                    };
-
-                    // If there is no manifest, create one
-                    let mut manifest = manifest.unwrap_or_else(|| {
+                        ManifestWriter::from_existing(
+                            &manifest_bytes,
+                            manifest,
+                            &manifest_schema,
+                            table_metadata,
+                            branch.as_deref(),
+                        )?
+                    } else {
                         let manifest_location = table_metadata.location.to_string()
                             + "/metadata/"
                             + snapshot_uuid
@@ -246,80 +234,20 @@ impl Operation {
                             + &0.to_string()
                             + ".avro";
 
-                        ManifestListEntry {
-                            format_version: table_metadata.format_version,
-                            manifest_path: manifest_location,
-                            manifest_length: 0,
-                            partition_spec_id: table_metadata.default_spec_id,
-                            content: Content::Data,
-                            sequence_number: table_metadata.last_sequence_number,
-                            min_sequence_number: 0,
-                            added_snapshot_id: snapshot_id,
-                            added_files_count: Some(0),
-                            existing_files_count: Some(0),
-                            deleted_files_count: Some(0),
-                            added_rows_count: Some(0),
-                            existing_rows_count: Some(0),
-                            deleted_rows_count: Some(0),
-                            partitions: None,
-                            key_metadata: None,
-                        }
-                    });
+                        ManifestWriter::new(
+                            &manifest_location,
+                            snapshot_id,
+                            &manifest_schema,
+                            table_metadata,
+                            branch.as_deref(),
+                        )?
+                    };
 
                     for manifest_entry in new_datafile_iter {
-                        {
-                            let manifest_entry = manifest_entry?;
-
-                            let mut added_rows_count = 0;
-
-                            if manifest.partitions.is_none() {
-                                manifest.partitions = Some(
-                                    table_metadata
-                                        .default_partition_spec()?
-                                        .fields()
-                                        .iter()
-                                        .map(|_| FieldSummary {
-                                            contains_null: false,
-                                            contains_nan: None,
-                                            lower_bound: None,
-                                            upper_bound: None,
-                                        })
-                                        .collect::<Vec<FieldSummary>>(),
-                                );
-                            }
-
-                            added_rows_count += manifest_entry.data_file().record_count();
-                            update_partitions(
-                                manifest.partitions.as_mut().unwrap(),
-                                manifest_entry.data_file().partition(),
-                                table_metadata.default_partition_spec()?.fields(),
-                            )?;
-
-                            manifest_writer.append_ser(manifest_entry)?;
-
-                            manifest.added_files_count = match manifest.added_files_count {
-                                Some(count) => Some(count + new_file_count as i32),
-                                None => Some(new_file_count as i32),
-                            };
-                            manifest.added_rows_count = match manifest.added_rows_count {
-                                Some(count) => Some(count + added_rows_count),
-                                None => Some(added_rows_count),
-                            };
-                        }
+                        manifest_writer.append(manifest_entry?)?;
                     }
 
-                    let manifest_bytes = manifest_writer.into_inner()?;
-
-                    let manifest_length: i64 = manifest_bytes.len() as i64;
-
-                    manifest.manifest_length += manifest_length;
-
-                    object_store
-                        .put(
-                            &strip_prefix(&manifest.manifest_path).as_str().into(),
-                            manifest_bytes.into(),
-                        )
-                        .await?;
+                    let manifest = manifest_writer.finish(object_store.clone()).await?;
 
                     manifest_list_writer.append_ser(manifest)?;
                 } else {
@@ -351,90 +279,26 @@ impl Operation {
                     };
 
                     for (i, entries) in splits.into_iter().enumerate() {
-                        let mut manifest_writer = ManifestWriter::new(
-                            Vec::new(),
-                            &manifest_schema,
-                            table_metadata,
-                            branch.as_deref(),
-                        )?;
-
                         let manifest_location = table_metadata.location.to_string()
                             + "/metadata/"
                             + snapshot_uuid
                             + "-m"
                             + &i.to_string()
                             + ".avro";
-                        let mut manifest = ManifestListEntry {
-                            format_version: table_metadata.format_version,
-                            manifest_path: manifest_location,
-                            manifest_length: 0,
-                            partition_spec_id: table_metadata.default_spec_id,
-                            content: Content::Data,
-                            sequence_number: table_metadata.last_sequence_number,
-                            min_sequence_number: 0,
-                            added_snapshot_id: snapshot_id,
-                            added_files_count: Some(0),
-                            existing_files_count: Some(0),
-                            deleted_files_count: Some(0),
-                            added_rows_count: Some(0),
-                            existing_rows_count: Some(0),
-                            deleted_rows_count: Some(0),
-                            partitions: None,
-                            key_metadata: None,
-                        };
+
+                        let mut manifest_writer = ManifestWriter::new(
+                            &manifest_location,
+                            snapshot_id,
+                            &manifest_schema,
+                            table_metadata,
+                            branch.as_deref(),
+                        )?;
 
                         for manifest_entry in entries {
-                            {
-                                let mut added_rows_count = 0;
-
-                                if manifest.partitions.is_none() {
-                                    manifest.partitions = Some(
-                                        table_metadata
-                                            .default_partition_spec()?
-                                            .fields()
-                                            .iter()
-                                            .map(|_| FieldSummary {
-                                                contains_null: false,
-                                                contains_nan: None,
-                                                lower_bound: None,
-                                                upper_bound: None,
-                                            })
-                                            .collect::<Vec<FieldSummary>>(),
-                                    );
-                                }
-
-                                added_rows_count += manifest_entry.data_file().record_count();
-                                update_partitions(
-                                    manifest.partitions.as_mut().unwrap(),
-                                    manifest_entry.data_file().partition(),
-                                    table_metadata.default_partition_spec()?.fields(),
-                                )?;
-
-                                manifest_writer.append_ser(manifest_entry)?;
-
-                                manifest.added_files_count = match manifest.added_files_count {
-                                    Some(count) => Some(count + new_file_count as i32),
-                                    None => Some(new_file_count as i32),
-                                };
-                                manifest.added_rows_count = match manifest.added_rows_count {
-                                    Some(count) => Some(count + added_rows_count),
-                                    None => Some(added_rows_count),
-                                };
-                            }
+                            manifest_writer.append(manifest_entry)?;
                         }
 
-                        let manifest_bytes = manifest_writer.into_inner()?;
-
-                        let manifest_length: i64 = manifest_bytes.len() as i64;
-
-                        manifest.manifest_length += manifest_length;
-
-                        object_store
-                            .put(
-                                &strip_prefix(&manifest.manifest_path).as_str().into(),
-                                manifest_bytes.into(),
-                            )
-                            .await?;
+                        let manifest = manifest_writer.finish(object_store.clone()).await?;
 
                         manifest_list_writer.append_ser(manifest)?;
                     }
@@ -564,12 +428,6 @@ impl Operation {
                 // Split manifest file if limit is exceeded
                 if n_splits == 0 {
                     // If manifest doesn't need to be split
-                    let mut manifest_writer = ManifestWriter::new(
-                        Vec::new(),
-                        &manifest_schema,
-                        table_metadata,
-                        branch.as_deref(),
-                    )?;
 
                     let manifest_location = table_metadata.location.to_string()
                         + "/metadata/"
@@ -577,79 +435,19 @@ impl Operation {
                         + "-m"
                         + &0.to_string()
                         + ".avro";
-                    let mut manifest = ManifestListEntry {
-                        format_version: table_metadata.format_version,
-                        manifest_path: manifest_location,
-                        manifest_length: 0,
-                        partition_spec_id: table_metadata.default_spec_id,
-                        content: Content::Data,
-                        sequence_number: table_metadata.last_sequence_number,
-                        min_sequence_number: 0,
-                        added_snapshot_id: snapshot_id,
-                        added_files_count: Some(0),
-                        existing_files_count: Some(0),
-                        deleted_files_count: Some(0),
-                        added_rows_count: Some(0),
-                        existing_rows_count: Some(0),
-                        deleted_rows_count: Some(0),
-                        partitions: None,
-                        key_metadata: None,
-                    };
+                    let mut manifest_writer = ManifestWriter::new(
+                        &manifest_location,
+                        snapshot_id,
+                        &manifest_schema,
+                        table_metadata,
+                        branch.as_deref(),
+                    )?;
 
                     for manifest_entry in new_datafile_iter {
-                        {
-                            let manifest_entry = manifest_entry?;
-
-                            let mut added_rows_count = 0;
-
-                            if manifest.partitions.is_none() {
-                                manifest.partitions = Some(
-                                    table_metadata
-                                        .default_partition_spec()?
-                                        .fields()
-                                        .iter()
-                                        .map(|_| FieldSummary {
-                                            contains_null: false,
-                                            contains_nan: None,
-                                            lower_bound: None,
-                                            upper_bound: None,
-                                        })
-                                        .collect::<Vec<FieldSummary>>(),
-                                );
-                            }
-
-                            added_rows_count += manifest_entry.data_file().record_count();
-                            update_partitions(
-                                manifest.partitions.as_mut().unwrap(),
-                                manifest_entry.data_file().partition(),
-                                table_metadata.default_partition_spec()?.fields(),
-                            )?;
-
-                            manifest_writer.append_ser(manifest_entry)?;
-
-                            manifest.added_files_count = match manifest.added_files_count {
-                                Some(count) => Some(count + new_file_count as i32),
-                                None => Some(new_file_count as i32),
-                            };
-                            manifest.added_rows_count = match manifest.added_rows_count {
-                                Some(count) => Some(count + added_rows_count),
-                                None => Some(added_rows_count),
-                            };
-                        }
+                        manifest_writer.append(manifest_entry?)?;
                     }
 
-                    let manifest_bytes = manifest_writer.into_inner()?;
-
-                    let manifest_length: i64 = manifest_bytes.len() as i64;
-
-                    manifest.manifest_length += manifest_length;
-
-                    object_store
-                        .put(
-                            &strip_prefix(&manifest.manifest_path).as_str().into(),
-                            manifest_bytes.into(),
-                        )
-                        .await?;
+                    let manifest = manifest_writer.finish(object_store.clone()).await?;
 
                     manifest_list_writer.append_ser(manifest)?;
                 } else {
@@ -662,90 +460,26 @@ impl Operation {
                     )?;
 
                     for (i, entries) in splits.into_iter().enumerate() {
-                        let mut manifest_writer = ManifestWriter::new(
-                            Vec::new(),
-                            &manifest_schema,
-                            table_metadata,
-                            branch.as_deref(),
-                        )?;
-
                         let manifest_location = table_metadata.location.to_string()
                             + "/metadata/"
                             + snapshot_uuid
                             + "-m"
                             + &i.to_string()
                             + ".avro";
-                        let mut manifest = ManifestListEntry {
-                            format_version: table_metadata.format_version,
-                            manifest_path: manifest_location,
-                            manifest_length: 0,
-                            partition_spec_id: table_metadata.default_spec_id,
-                            content: Content::Data,
-                            sequence_number: table_metadata.last_sequence_number,
-                            min_sequence_number: 0,
-                            added_snapshot_id: snapshot_id,
-                            added_files_count: Some(0),
-                            existing_files_count: Some(0),
-                            deleted_files_count: Some(0),
-                            added_rows_count: Some(0),
-                            existing_rows_count: Some(0),
-                            deleted_rows_count: Some(0),
-                            partitions: None,
-                            key_metadata: None,
-                        };
+
+                        let mut manifest_writer = ManifestWriter::new(
+                            &manifest_location,
+                            snapshot_id,
+                            &manifest_schema,
+                            table_metadata,
+                            branch.as_deref(),
+                        )?;
 
                         for manifest_entry in entries {
-                            {
-                                let mut added_rows_count = 0;
-
-                                if manifest.partitions.is_none() {
-                                    manifest.partitions = Some(
-                                        table_metadata
-                                            .default_partition_spec()?
-                                            .fields()
-                                            .iter()
-                                            .map(|_| FieldSummary {
-                                                contains_null: false,
-                                                contains_nan: None,
-                                                lower_bound: None,
-                                                upper_bound: None,
-                                            })
-                                            .collect::<Vec<FieldSummary>>(),
-                                    );
-                                }
-
-                                added_rows_count += manifest_entry.data_file().record_count();
-                                update_partitions(
-                                    manifest.partitions.as_mut().unwrap(),
-                                    manifest_entry.data_file().partition(),
-                                    table_metadata.default_partition_spec()?.fields(),
-                                )?;
-
-                                manifest_writer.append_ser(manifest_entry)?;
-
-                                manifest.added_files_count = match manifest.added_files_count {
-                                    Some(count) => Some(count + new_file_count as i32),
-                                    None => Some(new_file_count as i32),
-                                };
-                                manifest.added_rows_count = match manifest.added_rows_count {
-                                    Some(count) => Some(count + added_rows_count),
-                                    None => Some(added_rows_count),
-                                };
-                            }
+                            manifest_writer.append(manifest_entry)?;
                         }
 
-                        let manifest_bytes = manifest_writer.into_inner()?;
-
-                        let manifest_length: i64 = manifest_bytes.len() as i64;
-
-                        manifest.manifest_length += manifest_length;
-
-                        object_store
-                            .put(
-                                &strip_prefix(&manifest.manifest_path).as_str().into(),
-                                manifest_bytes.into(),
-                            )
-                            .await?;
+                        let manifest = manifest_writer.finish(object_store.clone()).await?;
 
                         manifest_list_writer.append_ser(manifest)?;
                     }
@@ -831,111 +565,4 @@ impl Operation {
             }
         }
     }
-}
-
-fn update_partitions(
-    partitions: &mut [FieldSummary],
-    partition_values: &Struct,
-    partition_columns: &[PartitionField],
-) -> Result<(), Error> {
-    for (field, summary) in partition_columns.iter().zip(partitions.iter_mut()) {
-        let value = partition_values.get(field.name()).and_then(|x| x.as_ref());
-        if let Some(value) = value {
-            if summary.lower_bound.is_none() {
-                summary.lower_bound = Some(value.clone());
-            } else if let Some(lower_bound) = &mut summary.lower_bound {
-                match (value, lower_bound) {
-                    (Value::Int(val), Value::Int(current)) => {
-                        if *current > *val {
-                            *current = *val
-                        }
-                    }
-                    (Value::LongInt(val), Value::LongInt(current)) => {
-                        if *current > *val {
-                            *current = *val
-                        }
-                    }
-                    (Value::Float(val), Value::Float(current)) => {
-                        if *current > *val {
-                            *current = *val
-                        }
-                    }
-                    (Value::Double(val), Value::Double(current)) => {
-                        if *current > *val {
-                            *current = *val
-                        }
-                    }
-                    (Value::Date(val), Value::Date(current)) => {
-                        if *current > *val {
-                            *current = *val
-                        }
-                    }
-                    (Value::Time(val), Value::Time(current)) => {
-                        if *current > *val {
-                            *current = *val
-                        }
-                    }
-                    (Value::Timestamp(val), Value::Timestamp(current)) => {
-                        if *current > *val {
-                            *current = *val
-                        }
-                    }
-                    (Value::TimestampTZ(val), Value::TimestampTZ(current)) => {
-                        if *current > *val {
-                            *current = *val
-                        }
-                    }
-                    _ => {}
-                }
-            }
-            if summary.upper_bound.is_none() {
-                summary.upper_bound = Some(value.clone());
-            } else if let Some(upper_bound) = &mut summary.upper_bound {
-                match (value, upper_bound) {
-                    (Value::Int(val), Value::Int(current)) => {
-                        if *current < *val {
-                            *current = *val
-                        }
-                    }
-                    (Value::LongInt(val), Value::LongInt(current)) => {
-                        if *current < *val {
-                            *current = *val
-                        }
-                    }
-                    (Value::Float(val), Value::Float(current)) => {
-                        if *current < *val {
-                            *current = *val
-                        }
-                    }
-                    (Value::Double(val), Value::Double(current)) => {
-                        if *current < *val {
-                            *current = *val
-                        }
-                    }
-                    (Value::Date(val), Value::Date(current)) => {
-                        if *current < *val {
-                            *current = *val
-                        }
-                    }
-                    (Value::Time(val), Value::Time(current)) => {
-                        if *current < *val {
-                            *current = *val
-                        }
-                    }
-                    (Value::Timestamp(val), Value::Timestamp(current)) => {
-                        if *current < *val {
-                            *current = *val
-                        }
-                    }
-                    (Value::TimestampTZ(val), Value::TimestampTZ(current)) => {
-                        if *current < *val {
-                            *current = *val
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-    }
-    Ok(())
 }
